+++ cliquet/resource/__init__.py
@@ -179,6 +179,15 @@ class UserResource(object):
         """
         return getattr(request, 'prefixed_userid', None)
 
+    def _get_known_fields(self):
+        """Return all the `field` defined in the ressource mapping."""
+        known_fields = [c.name for c in self.mapping.children] + \
+                       [self.model.id_field,
+                        self.model.modified_field,
+                        self.model.deleted_field]
+        return known_fields
+
+
     def is_known_field(self, field):
         """Return ``True`` if `field` is defined in the resource mapping.
 
@@ -186,10 +195,7 @@ class UserResource(object):
         :rtype: bool
 
         """
-        known_fields = [c.name for c in self.mapping.children] + \
-                       [self.model.id_field,
-                        self.model.modified_field,
-                        self.model.deleted_field]
+        known_fields = self._get_known_fields()
         return field in known_fields
 
     #
@@ -219,6 +225,8 @@ class UserResource(object):
         filters = self._extract_filters()
         limit = self._extract_limit()
         sorting = self._extract_sorting(limit)
+        partial_fields = self._extract_fields()
+
         filter_fields = [f.field for f in filters]
         include_deleted = self.model.modified_field in filter_fields
 
@@ -239,6 +247,12 @@ class UserResource(object):
             lastrecord = records[-1]
             next_page = self._next_page_url(sorting, limit, lastrecord, offset)
             headers['Next-Page'] = encode_header(next_page)
+
+        if partial_fields:
+            records = [
+                self._do_projection(record, partial_fields)
+                for record in records
+            ]
 
         # Bind metric about response size.
         logger.bind(nb_records=len(records), limit=limit)
@@ -329,6 +343,10 @@ class UserResource(object):
         self._add_cache_header(self.request.response)
         self._raise_304_if_not_modified(record)
         self._raise_412_if_modified(record)
+
+        partial_fields = self._extract_fields()
+        if partial_fields:
+            record = self._do_projection(record, partial_fields)
 
         return self.postprocess(record)
 
@@ -826,6 +844,31 @@ class UserResource(object):
                 'description': error_msg
             }
             raise_invalid(self.request, **error_details)
+
+    def _do_projection(self, record, partial_fields):
+        """Project record on partial_fields."""
+
+        return {field: record.get(field) for field in partial_fields}
+
+    def _extract_fields(self):
+        """Extract the fields to do the projection from QueryString parameters.
+        """
+
+        fields = self.request.GET.get('_fields', None)
+        if fields:
+            fields = fields.split(',')
+            known_fields = self._get_known_fields()
+            invalid_fields = set(fields) - set(known_fields)
+            if invalid_fields:
+                error_msg = "Fields {} do not exist".\
+                    format(set(fields) - set(known_fields))
+                error_details = {
+                    'name': "Invalid _fields parameter",
+                    'description': error_msg
+                }
+                raise_invalid(self.request, **error_details)
+
+        return fields
 
     def _extract_limit(self):
         """Extract limit value from QueryString parameters."""